+++ extracter/jsonpath_test.go
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package jsonextracter
+package extracter
 
 import (
 	"encoding/json"
